feat(product): add confirmation dialog for product deletion
Enhanced the deleteProduct method in the product card component to include a confirmation dialog before proceeding with the deletion. Added a notification for action cancellation to improve user experience.

+++ src/app/pages/catalog/product/product-card/product-card.component.ts
@@ -207,20 +207,24 @@ export class ProductCardComponent implements OnInit {
   }
 
   deleteProduct() {
-    this.ProductHeadService.deleteProductHead(this.headDetails?._id).subscribe({
-      next: (res: any) => {
-        if (res?.errorCode == 0) {
-          this.ToastrService.success(res?.message);
-          this.close();
-          this.getProductHeads();
-        } else {
-          this.ToastrService.error(res?.message);
-        }
-      },
-      error: (err: any) => {
-        this.ToastrService.error(err?.error?.message);
-      },
-    });
+    if (confirm('Are you sure you want to delete this product?')) {
+      this.ProductHeadService.deleteProductHead(this.headDetails?._id).subscribe({
+        next: (res: any) => {
+          if (res?.errorCode == 0) {
+            this.ToastrService.success(res?.message);
+            this.close();
+            this.getProductHeads();
+          } else {
+            this.ToastrService.error(res?.message);
+          }
+        },
+        error: (err: any) => {
+          this.ToastrService.error(err?.error?.message);
+        },
+      });
+    } else {
+      this.HotToastService.error('Action cancelled');
+    }
   }
 
   ngOnInit(): void {
